Changes
Set dressing for the white box is complete. Adjusted the crate sizes and made a small change in room 2 for the crate layout on how the player can move to and from the vents. Added the guard to the end of the room, followed by the doors hovering above the frames so they could be used and played with